Split todo controller into parts
