+++ AdvancedTodoList/Controllers/TodoItemsController.cs
@@ -9,87 +9,11 @@ namespace AdvancedTodoList.Controllers;
 
 [Authorize]
 [ApiController]
-[Route("api/todo")]
-public class TodoController(
-	ITodoListsService todoListsService, ITodoItemsService todoItemsService,
-	IEntityExistenceChecker existenceChecker) : ControllerBase
+[Route("api/todo/{listId}/items")]
+public class TodoItemsController(ITodoItemsService todoItemsService) : ControllerBase
 {
-	private readonly ITodoListsService _todoListsService = todoListsService;
 	private readonly ITodoItemsService _todoItemsService = todoItemsService;
-	private readonly IEntityExistenceChecker _existenceChecker = existenceChecker;
 
-	#region Lists
-	/// <summary>
-	/// Gets a to-do list by its ID.
-	/// </summary>
-	/// <param name="listId">ID of the to-do list to obtain.</param>
-	/// <response code="200">Returns requested to-do list.</response>
-	/// <response code="401">Authentication failed.</response>
-	/// <response code="404">To-do list was not found.</response>
-	[HttpGet("{listId}", Name = nameof(GetTodoListByIdAsync))]
-	[ProducesResponseType(typeof(TodoListGetByIdDto), StatusCodes.Status200OK)]
-	[ProducesResponseType(StatusCodes.Status401Unauthorized)]
-	[ProducesResponseType(StatusCodes.Status404NotFound)]
-	public async Task<IActionResult> GetTodoListByIdAsync([FromRoute] string listId)
-	{
-		var list = await _todoListsService.GetByIdAsync(listId);
-		return list != null ? Ok(list) : NotFound();
-	}
-
-	/// <summary>
-	/// Creates a new to-do list.
-	/// </summary>
-	/// <response code="201">Successfully created.</response>
-	/// <response code="400">Validation failed.</response>
-	/// <response code="401">Authentication failed.</response>
-	[HttpPost]
-	[ProducesResponseType(typeof(TodoListGetByIdDto), StatusCodes.Status201Created)]
-	[ProducesResponseType(StatusCodes.Status400BadRequest)]
-	[ProducesResponseType(StatusCodes.Status401Unauthorized)]
-	public async Task<IActionResult> PostTodoListAsync([FromBody] TodoListCreateDto dto)
-	{
-		var list = await _todoListsService.CreateAsync(dto);
-		var routeValues = new { listId = list.Id };
-		return CreatedAtRoute(nameof(GetTodoListByIdAsync), routeValues, list);
-	}
-
-	/// <summary>
-	/// Updates a to-do list.
-	/// </summary>
-	/// <response code="204">Success.</response>
-	/// <response code="400">Validation failed.</response>
-	/// <response code="401">Authentication failed.</response>
-	/// <response code="404">To-do list was not found.</response>
-	[HttpPut("{listId}")]
-	[ProducesResponseType(StatusCodes.Status204NoContent)]
-	[ProducesResponseType(StatusCodes.Status400BadRequest)]
-	[ProducesResponseType(StatusCodes.Status401Unauthorized)]
-	[ProducesResponseType(StatusCodes.Status404NotFound)]
-	public async Task<IActionResult> PutTodoListAsync(
-		[FromRoute] string listId, [FromBody] TodoListCreateDto dto)
-	{
-		bool result = await _todoListsService.EditAsync(listId, dto);
-		return result ? NoContent() : NotFound();
-	}
-
-	/// <summary>
-	/// Deletes a to-do list.
-	/// </summary>
-	/// <response code="204">Success.</response>
-	/// <response code="401">Authentication failed.</response>
-	/// <response code="404">To-do list was not found.</response>
-	[HttpDelete("{listId}")]
-	[ProducesResponseType(StatusCodes.Status204NoContent)]
-	[ProducesResponseType(StatusCodes.Status401Unauthorized)]
-	[ProducesResponseType(StatusCodes.Status404NotFound)]
-	public async Task<IActionResult> DeleteTodoListAsync([FromRoute] string listId)
-	{
-		bool result = await _todoListsService.DeleteAsync(listId);
-		return result ? NoContent() : NotFound();
-	}
-	#endregion
-
-	#region Items
 	/// <summary>
 	/// Gets a page with items of the to-do list with the specified ID.
 	/// </summary>
@@ -98,7 +22,7 @@ public class TodoController(
 	/// <response code="200">Returns items of the to-do list.</response>
 	/// <response code="401">Authentication failed.</response>
 	/// <response code="404">To-do list was not found.</response>
-	[HttpGet("{listId}/items", Name = nameof(GetListItemsAsync))]
+	[HttpGet(Name = nameof(GetListItemsAsync))]
 	[ProducesResponseType(typeof(Page<TodoItemPreviewDto>), StatusCodes.Status200OK)]
 	[ProducesResponseType(StatusCodes.Status401Unauthorized)]
 	[ProducesResponseType(StatusCodes.Status404NotFound)]
@@ -118,7 +42,7 @@ public class TodoController(
 	/// <response code="200">Returns requested to-do list item.</response>
 	/// <response code="401">Authentication failed.</response>
 	/// <response code="404">To-do list item was not found.</response>
-	[HttpGet("{listId}/items/{itemId}", Name = nameof(GetTodoItemByIdAsync))]
+	[HttpGet("{itemId}", Name = nameof(GetTodoItemByIdAsync))]
 	[ProducesResponseType(typeof(TodoItemGetByIdDto), StatusCodes.Status200OK)]
 	[ProducesResponseType(StatusCodes.Status401Unauthorized)]
 	[ProducesResponseType(StatusCodes.Status404NotFound)]
@@ -141,7 +65,7 @@ public class TodoController(
 	/// <response code="400">Validation failed.</response>
 	/// <response code="401">Authentication failed.</response>
 	/// <response code="404">To-do list was not found.</response>
-	[HttpPost("{listId}/items")]
+	[HttpPost]
 	[ProducesResponseType(typeof(TodoItemGetByIdDto), StatusCodes.Status201Created)]
 	[ProducesResponseType(StatusCodes.Status400BadRequest)]
 	[ProducesResponseType(StatusCodes.Status401Unauthorized)]
@@ -164,7 +88,7 @@ public class TodoController(
 	/// <response code="400">Validation failed.</response>
 	/// <response code="401">Authentication failed.</response>
 	/// <response code="404">To-do list item was not found.</response>
-	[HttpPut("{listId}/items/{itemId}")]
+	[HttpPut("{itemId}")]
 	[ProducesResponseType(StatusCodes.Status204NoContent)]
 	[ProducesResponseType(StatusCodes.Status400BadRequest)]
 	[ProducesResponseType(StatusCodes.Status401Unauthorized)]
@@ -184,7 +108,7 @@ public class TodoController(
 	/// <response code="400">Validation failed.</response>
 	/// <response code="401">Authentication failed.</response>
 	/// <response code="404">To-do list item was not found.</response>
-	[HttpPut("{listId}/items/{itemId}/state")]
+	[HttpPut("{itemId}/state")]
 	[ProducesResponseType(StatusCodes.Status204NoContent)]
 	[ProducesResponseType(StatusCodes.Status400BadRequest)]
 	[ProducesResponseType(StatusCodes.Status401Unauthorized)]
@@ -203,7 +127,7 @@ public class TodoController(
 	/// <response code="204">Success.</response>
 	/// <response code="401">Authentication failed.</response>
 	/// <response code="404">To-do list item was not found.</response>
-	[HttpDelete("{listId}/items/{itemId}")]
+	[HttpDelete("{itemId}")]
 	[ProducesResponseType(StatusCodes.Status204NoContent)]
 	[ProducesResponseType(StatusCodes.Status401Unauthorized)]
 	[ProducesResponseType(StatusCodes.Status404NotFound)]
@@ -213,5 +137,4 @@ public class TodoController(
 		bool result = await _todoItemsService.DeleteAsync(listId, itemId);
 		return result ? NoContent() : NotFound();
 	}
-	#endregion
 }
